Adding optional push buttons for local control
Going beyond the spec. I've added a few lines of code to enable the use of push buttons as well as DCC control. Adding push buttons means that you can activate the uncoupler without having to use the DCC handset. One leg of a push button is connected to an Arduino I/O pin, the other leg to GND. The Arduino's internal pullup resistors are enabled for the push button pins. This means the input is normally HIGH and goes LOW when the button is pressed.